+++ src/core/search/scorer/mod.rs
@@ -47,6 +47,10 @@ mod boosting_scorer;
 
 pub use self::boosting_scorer::*;
 
+mod phrase_scorer;
+
+pub use self::phrase_scorer::*;
+
 use std::collections::HashMap;
 use std::i32;
